requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.2
itsdangerous==2.2.0
JayDeBeApi==1.2.3
Jinja2==3.1.6
JPype1==1.4.1
MarkupSafe==3.0.2
packaging==24.2
prometheus-client==0.17.1
pycparser==2.22
pyOpenSSL==24.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
Werkzeug==3.1.3
